+++ BackEnd_TrecoLista/Infraestrutura/Util/ValorConverter.cs
@@ -4,7 +4,7 @@ namespace BackEnd_TrecoLista.Infraestrutura.Util
 {
     public static class ValorConverter
     {
-        public static double ConvertPrice(string preco)
+        public static decimal ConvertPrice(string preco)
         {
             // Remove todos os caracteres não numéricos, exceto o ponto e a vírgula
             var cleanedPrice = new string(preco.Where(c => char.IsDigit(c) || c == '.' || c == ',').ToArray());
@@ -13,7 +13,7 @@ namespace BackEnd_TrecoLista.Infraestrutura.Util
             cleanedPrice = cleanedPrice.Replace(',', '.');
 
             // Converter para double
-            if (double.TryParse(cleanedPrice, NumberStyles.Any, CultureInfo.InvariantCulture, out double result))
+            if (decimal.TryParse(cleanedPrice, NumberStyles.Any, CultureInfo.InvariantCulture, out decimal result))
             {
                 return result;
             }
